#11269 Fix Supplier Return delete hanging on detail page
Returning the invalidateQueries promise from onSuccess made
mutateAsync await a refetch of the just-deleted detail query.
With React Query retry enabled (production config), that retries
forever and the confirmation modal's OK button never resolves.

Use `void` to make the fire-and-forget intent explicit and prevent
the awaited refetch. Apply the same shape to the customer return
hook for consistency, and move that hook's navigation into the
sidepanel to keep the hook route-agnostic.
