+++ client/packages/invoices/src/Returns/api/hooks/document/useCustomerDelete.ts
@@ -1,25 +1,13 @@
-import {
-  RouteBuilder,
-  useMutation,
-  useNavigate,
-  useQueryClient,
-} from '@openmsupply-client/common';
-import { AppRoute } from '@openmsupply-client/config';
+import { useMutation, useQueryClient } from '@openmsupply-client/common';
 import { useReturnsApi } from '../utils/useReturnsApi';
 
 export const useCustomerReturnDelete = () => {
   const queryClient = useQueryClient();
   const api = useReturnsApi();
-  const navigate = useNavigate();
 
   return useMutation(api.deleteCustomer, {
-    onSuccess: () => {
-      queryClient.invalidateQueries(api.keys.base());
-      navigate(
-        RouteBuilder.create(AppRoute.Distribution)
-          .addPart(AppRoute.CustomerReturn)
-          .build()
-      );
-    },
+    // `void` is load-bearing: returning the invalidateQueries promise would make
+    // mutateAsync await the refetch of the just-deleted detail query and hang.
+    onSuccess: () => void queryClient.invalidateQueries(api.keys.base()),
   });
 };
